LUC065-364

+++ dspace-api/src/main/java/org/dspace/eperson/Groomer.java
@@ -141,10 +141,20 @@ public class Groomer {
             System.out.println();
 
             if (delete) {
-                try {
-                    ePersonService.delete(myContext, account);
-                } catch (AuthorizeException | IOException ex) {
-                    System.err.println(ex.getMessage());
+                List<String> whyNot = ePersonService.getDeleteConstraints(myContext, account);
+                if (!whyNot.isEmpty()) {
+                    System.out.print("\tCannot be deleted; referenced in");
+                    for (String table : whyNot) {
+                        System.out.print(' ');
+                        System.out.print(table);
+                    }
+                    System.out.println();
+                } else {
+                    try {
+                        ePersonService.delete(myContext, account);
+                    } catch (AuthorizeException | IOException ex) {
+                        System.err.println(ex.getMessage());
+                    }
                 }
             }
         }
